refactor(github): extract GitHub tree hash logic
