+++ src/skill-lock.ts
@@ -2,18 +2,10 @@ import { readFile, writeFile, mkdir } from 'fs/promises';
 import { join, dirname } from 'path';
 import { homedir } from 'os';
 import { createHash } from 'crypto';
-import { execSync } from 'child_process';
 
 const AGENTS_DIR = '.agents';
 const LOCK_FILE = '.skill-lock.json';
 const CURRENT_VERSION = 3; // Bumped from 2 to 3 for folder hash support (GitHub tree SHA)
-const repoTreeCache = new Map<
-  string,
-  Promise<{
-    sha: string;
-    tree: Array<{ path: string; type: string; sha: string }>;
-  } | null>
->();
 
 /**
  * Represents a single installed skill entry in the lock file.
@@ -126,128 +118,6 @@ export async function writeSkillLock(lock: SkillLockFile): Promise<void> {
  */
 export function computeContentHash(content: string): string {
   return createHash('sha256').update(content, 'utf-8').digest('hex');
-}
-
-/**
- * Get GitHub token from user's environment.
- * Tries in order:
- * 1. GITHUB_TOKEN environment variable
- * 2. GH_TOKEN environment variable
- * 3. gh CLI auth token (if gh is installed)
- *
- * @returns The token string or null if not available
- */
-export function getGitHubToken(): string | null {
-  // Check environment variables first
-  if (process.env.GITHUB_TOKEN) {
-    return process.env.GITHUB_TOKEN;
-  }
-  if (process.env.GH_TOKEN) {
-    return process.env.GH_TOKEN;
-  }
-
-  // Try gh CLI
-  try {
-    const token = execSync('gh auth token', {
-      encoding: 'utf-8',
-      stdio: ['pipe', 'pipe', 'pipe'],
-    }).trim();
-    if (token) {
-      return token;
-    }
-  } catch {
-    // gh not installed or not authenticated
-  }
-
-  return null;
-}
-
-/**
- * Fetch the tree SHA (folder hash) for a skill folder using GitHub's Trees API.
- * This makes ONE API call to get the entire repo tree, then extracts the SHA
- * for the specific skill folder.
- *
- * @param ownerRepo - GitHub owner/repo (e.g., "vercel-labs/agent-skills")
- * @param skillPath - Path to skill folder or SKILL.md (e.g., "skills/react-best-practices/SKILL.md")
- * @param token - Optional GitHub token for authenticated requests (higher rate limits)
- * @returns The tree SHA for the skill folder, or null if not found
- */
-export async function fetchSkillFolderHash(
-  ownerRepo: string,
-  skillPath: string,
-  token?: string | null
-): Promise<string | null> {
-  // Normalize to forward slashes first (for GitHub API compatibility)
-  let folderPath = skillPath.replace(/\\/g, '/');
-
-  // Remove SKILL.md suffix to get folder path
-  if (folderPath.endsWith('/SKILL.md')) {
-    folderPath = folderPath.slice(0, -9);
-  } else if (folderPath.endsWith('SKILL.md')) {
-    folderPath = folderPath.slice(0, -8);
-  }
-
-  // Remove trailing slash
-  if (folderPath.endsWith('/')) {
-    folderPath = folderPath.slice(0, -1);
-  }
-
-  const branches = ['main', 'master'];
-
-  for (const branch of branches) {
-    const cacheKey = `${ownerRepo}@${branch}`;
-    let treePromise = repoTreeCache.get(cacheKey);
-
-    if (!treePromise) {
-      treePromise = (async () => {
-        const url = `https://api.github.com/repos/${ownerRepo}/git/trees/${branch}?recursive=1`;
-        const headers: Record<string, string> = {
-          Accept: 'application/vnd.github.v3+json',
-          'User-Agent': 'skills-cli',
-        };
-        if (token) {
-          headers['Authorization'] = `Bearer ${token}`;
-        }
-
-        const response = await fetch(url, { headers });
-
-        if (!response.ok) return null;
-
-        return (await response.json()) as {
-          sha: string;
-          tree: Array<{ path: string; type: string; sha: string }>;
-        };
-      })();
-      repoTreeCache.set(cacheKey, treePromise);
-    }
-
-    try {
-      const data = await treePromise;
-      if (!data) {
-        repoTreeCache.delete(cacheKey);
-        continue;
-      }
-
-      // If folderPath is empty, this is a root-level skill - use the root tree SHA
-      if (!folderPath) {
-        return data.sha;
-      }
-
-      // Find the tree entry for the skill folder
-      const folderEntry = data.tree.find(
-        (entry) => entry.type === 'tree' && entry.path === folderPath
-      );
-
-      if (folderEntry) {
-        return folderEntry.sha;
-      }
-    } catch {
-      repoTreeCache.delete(cacheKey);
-      continue;
-    }
-  }
-
-  return null;
 }
 
 /**
